requirements.txt
keras==2.11.0
numpy==1.24.2
Pillow==9.4.0
streamlit==1.20.0
tensorflow==2.11.0
tensorflow_intel==2.11.0
